Moved files to helpers folder
